integrate gRPC TrendingMoviesService client into API gateway and update catalog aggregation logic

+++ movie-catalog-service/src/main/java/com/moviecatalogservice/models/CatalogItem.java
@@ -4,13 +4,13 @@ public class CatalogItem {
 
     private String name;
     private String description;
-    private int rating;
+    private double rating;
 
     public CatalogItem() {
 
     }
 
-    public CatalogItem(String name, String description, int rating) {
+    public CatalogItem(String name, String description, double rating) {
         this.name = name;
         this.description = description;
         this.rating = rating;
@@ -32,11 +32,11 @@ public class CatalogItem {
         this.description = description;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
 }
